// src/data/types.ts
export type ClientId = 'oi' | 'ku'
// ...
export interface PolicyCitation {
  researcher: string
  client: string
  topic: string
  coreTopic: boolean
  citedPaperTitle: string
  citedPaperYear: number
  citedPaperDOI: string
  policyDocTitle: string
  policyDocYear: number
  policyDocType: string
  citingOrg: string
  country: string
  orgType: string
  url: string
}